stagefour/requirements.txt
# Gateway + 现有 Agent 依赖（按需与项目其他部分对齐）
python-dotenv>=1.0.0
openai>=1.40.0
mcp>=1.0.0
httpx>=0.27.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
